ac || argv

+++ philo.h
@@ -3,14 +3,15 @@
 
 # include <stdio.h>
 # include <pthread.h>
+# include <limits.h>
 
 typedef struct s_data
 {
-    int n_philos;
-    int time_to_die;
-    int time_to_eat;
-    int time_to_sleep;
-    int n_eat;
+    long    n_philos;
+    long    time_to_die;
+    long    time_to_eat;
+    long    time_to_sleep;
+    int     n_eat;
 }               t_data;
